EPGsim.EPGStates

EPGsim.EPGStates

EPGsim.epgDephasing!

EPGsim.epgRelaxation!

EPGsim.epgRotation!

EPGsim.epgRotation!

EPGsim.epgSignal

EPGsim.rfRotation

EPGsim.EPGStates — Type

EPGStates{T <: Real}

Stores the EPG states in 3 vectors Fp,Fn and Z.

Constructors :

EPGStates(Fp::Vector{Complex{S}},Fn::Vector{Complex{S}},Z::Vector{Complex{S}}) where {S <: Real}
EPGStates(Fp::T=0,Fn::T=0,Z::T=1) where T <: Number

Fields

Fp::Vector{Complex{T}}

Fn::Vector{Complex{T}}

Z::Vector{Complex{T}}

Related functions

getStates(E::EPGStates) : extract EPG states as matrix 3xN

EPGsim.EPGStates — Method

getStates(E::EPGStates)

Extract EPG states as matrix 3xN

EPGsim.epgDephasing! — Function

epgDephasing!(E::EPGStates, n=1) where T

shifts the transverse dephasing states F corresponding to n dephasing-cycles. n can be any integer

EPGsim.epgRelaxation! — Method

epgRelaxation!(E::EPGStates,t,T1, T2)

applies relaxation matrices to a set of EPG states.

Arguments

E::EPGStates

t::AbstractFloat - length of time interval

T1::AbstractFloat - T1

T2::AbstractFloat - T2

EPGsim.epgRotation! — Function

epgRotation!(E::EPGStates, alpha::Float64, phi::Float64=0.0)

applies Bloch-rotation (<=> RF pulse) to a set of EPG states.

Arguments

E::EPGStates`

alpha::Float64 - flip angle of the RF pulse (rad)

phi::Float64=0.0 - phase of the RF pulse (rad)

EPGsim.epgRotation! — Method

epgRotation!(E::EPGStates, R::Matrix)

applies rotation matrix from rfRotation function to the EPGStates

Arguments

E::EPGStates`

R::Matrix - rotation Matrix (rad)

EPGsim.epgSignal — Function

epgSignal(E::EPGStates, phase::Real=0.0)

Return the signal corresponding to E.Fp[1]. If phase is provided (radians), the returned value is corrected by multiplying by exp(-im*phase).

EPGsim.rfRotation — Function

rfRotation(alpha, phi=0.)

returns the rotation matrix for a pulse with flip angle alpha and phase phi.

Arguments

alpha - flip angle (radian)

phi=0. - phase of the flip angle (radian)