EPGsim.epgDephasing

EPGsim.epgRelaxation

EPGsim.epgRotation

EPGsim.rfRotation

EPGsim.epgDephasing — Function

epgDephasing(E::EPGStates, n=1) where T

shifts the transverse dephasing states F corresponding to n dephasing-cycles.

EPGsim.epgRelaxation — Method

epgRelaxation(E::EPGStates,t,T1, T2)

applies relaxation matrices to a set of EPG states.

Arguments

E::EPGStates

t::AbstractFloat - length of time interval

T1::AbstractFloat - T1

T2::AbstractFloat - T2

EPGsim.epgRotation — Function

epgRotation(E::EPGStates, alpha::Float64, phi::Float64=0.0)

applies Bloch-rotation (<=> RF pulse) to a set of EPG states.

Arguments

E::EPGStates`

alpha::Float64 - flip angle of the RF pulse

phi::Float64=0.0 - phase of the RF pulse

EPGsim.rfRotation — Function

rfRotation_AT(alpha, phi=0.)

returns the rotation matrix for a pulse with flip angle alpha and phase phi.